% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barras_horizontales.R
\name{barras_horizontales}
\alias{barras_horizontales}
\title{Title}
\usage{
barras_horizontales(df, valor, cat, nudge_text = 0, espacio_extra = 0.15)
}
\arguments{
\item{espacio_extra}{}
}
\value{

}
\description{
Title
}
