% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barras_horizontales.R
\name{barras_horizontales_2}
\alias{barras_horizontales_2}
\title{Barras Horizonales factor ordernado}
\usage{
barras_horizontales_2(
  df,
  valor,
  cat,
  nudge = 0.12,
  accuracy = 1,
  espacio_extra = 0.15
)
}
\arguments{
\item{cat}{una columna de df que es un factor ordenado}

\item{espacio_extra}{}
}
\description{
Barras Horizonales factor ordernado
}
