% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pie_chart.R
\name{pie_chart}
\alias{pie_chart}
\title{Make pie chart}
\usage{
pie_chart(
  df,
  values,
  labels,
  nudge_radio = 0.7,
  nudge_x = 0,
  nudge_y = 0,
  percent_accuracy = 0.1
)
}
\arguments{
\item{df}{data_frame}

\item{values}{variable a graficar}

\item{labels}{variable para agrupar}

\item{nudge_radio}{distancia del centro para las etiquetas}

\item{nudge_x}{vector para mover las  etiquetas en eje x}

\item{nudge_y}{vector para mover las etiquetas en eje  y}
}
\value{

}
\description{
Crear grafico de tortas
}
\examples{



}
