% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serie_historica.R
\name{serie_historica}
\alias{serie_historica}
\title{Serie Histórica}
\usage{
serie_historica(df, valor, fecha)
}
\arguments{
\item{df}{data frame}

\item{valor}{valor a graficar}

\item{fecha}{fecha}
}
\value{
grafico
}
\description{
Serie Histórica
}
\examples{



}
