% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serie_historica.R
\name{serie_historica_anual}
\alias{serie_historica_anual}
\title{Serie Historica Anual}
\usage{
serie_historica_anual(df, valor)
}
\arguments{
\item{df}{datos}

\item{valor}{valor a graficar}
}
\value{

}
\description{
Serie Historica Anual
}
