% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serie_historica.R
\name{serie_historica_mensual}
\alias{serie_historica_mensual}
\title{Title}
\usage{
serie_historica_mensual(df, valor, col_width = 0.5)
}
\arguments{
\item{df}{datos}

\item{valor}{valores de la serie}

\item{col_width}{ancho de las columnas}
}
\value{

}
\description{
Title
}
