% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serie_historica.R
\name{serie_historica_semestral_flujo}
\alias{serie_historica_semestral_flujo}
\title{Title}
\usage{
serie_historica_semestral_flujo(df, valor)
}
\arguments{
\item{df}{datos}

\item{valor}{a graficar}
}
\description{
Title
}
