% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barras_horizontales.R
\name{barras_horizontales}
\alias{barras_horizontales}
\title{Title}
\usage{
barras_horizontales(
  df,
  valor,
  cat,
  nudge = 0.12,
  accuracy = 1,
  espacio_extra = 0.15
)
}
\arguments{
\item{df}{datos}

\item{valor}{columna con los valores a graficar}

\item{cat}{columnas con las categorias}

\item{nudge}{valor para mover el texto}

\item{accuracy}{precision para los numeros (no los porcentajes!) de las etiquetas}

\item{espacio_extra}{espacio}
}
\value{
plot object
}
\description{
Title
}
