% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serie_historica.R
\name{serie_historica}
\alias{serie_historica}
\title{Serie Histórica}
\usage{
serie_historica(df, valor, fecha, nudge_text = 0, size_text = 5)
}
\value{

}
\description{
Serie Histórica
}
\examples{



}
