% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serie_historica.R
\name{serie_historica_anual}
\alias{serie_historica_anual}
\title{Serie Historica Anual}
\usage{
serie_historica_anual(df, valor, nudge_text = 0, size_text = 5)
}
\arguments{
\item{size_text}{}
}
\value{

}
\description{
Serie Historica Anual
}
