% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serie_historica.R
\name{serie_historica_mensual}
\alias{serie_historica_mensual}
\title{Title}
\usage{
serie_historica_mensual(
  df,
  valor,
  nudge_text = 0,
  size_text = 5,
  col_width = 0.5
)
}
\arguments{
\item{col_width}{}
}
\value{

}
\description{
Title
}
