% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serie_historica.R
\name{serie_historica_semestral_flujo}
\alias{serie_historica_semestral_flujo}
\title{Title}
\usage{
serie_historica_semestral_flujo(df, valor, nudge_text = 0, size_text = 5)
}
\arguments{
\item{col_width}{}
}
\value{

}
\description{
Title
}
