% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serie_historica.R
\name{serie_historica_semestral_stock}
\alias{serie_historica_semestral_stock}
\title{Title}
\usage{
serie_historica_semestral_stock(df, valor, nudge_text = 0)
}
\arguments{
\item{nudge_text}{}
}
\value{

}
\description{
Title
}
